requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.1
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-ipware==7.0.1
django-redis==6.0.0
django-rest-authemail==2.1.7
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
python-ipware==3.0.0
redis==6.2.0
requests==2.32.4
rq==2.4.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
